tools/requirements.txt
xrpl-py>=2.4.0
